Improve uptime of edge inference container + small fixes (#50)
I was setting up an edge endpoint from scratch this weekend and realized a few issues and opportunities for improvement.

The main thing that this change does is it moves all model updating (including management of Kubernetes inference deployments) to the inference-model-updater container. I think this is a lot cleaner than doing work of this nature in the image_queries view.

Additionally, I modified the inference k8s config to avoid downtime during updates by adding a rollout strategy and adding a readinessProbe to ensure the model is loaded and ready to go before requests are sent to it (and before the previous inference pod is terminated).

I also fixed an issue where we never created an initial inference deployment (until we had a new model version).