Add timestamps to inference-model-updater logs (#106)
Previously, logs in the inference-model-updater container of the
edge-endpoint pod didn't have timestamps. Now they do! This makes it
easier to know when the next update is going to happen, and generally to
debug model updating.

I'm not an expert on how python logging setup works, so let me know if
this should be done in a different way.