Add rate limit to background escalations (#115)
This PR adds a rate limit to detectors running background escalations to
the cloud. The default time to wait between escalations will be 2
seconds, but we'll allow customers to set that for themselves. If they
set too small of a limit (or a limit of 0), they'll still be effectively
limited by our usual throttling.

Unit test coverage of `image_queries.py` is still a little lacking, and
that's a larger fix than I think makes sense to do in this PR. I tested
out the new option with the demo code.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>