Rewording things so we use "cluster" in the canonical sense of the word (#118)
Before this, our docs & code would use the word "cluster" to sometimes
mean our application, sometimes mean the namespace, and sometimes to
actually mean the k8 cluster. I think now whenever we use the word
cluster we're doing so in a way that would make sense to anybody versed
with k8.