Update edge submodule in glhub on merges to main (#145)
I've created a bot account that can push directly to glhub main
(bypasses branch protection rule). We run this action in our protected
`live` environment that me, Harry, Brandon, Corey, Tom and Tyler have
access to. We approve deployments to this environment the same way we do
for zuuul. Once the deployment is approved, the action will update glhub
main. We've set it up this way to protect the bot account's github
access token -- it lives only within the protected live environment.
This action runs on merges to main of edge endpoint.