Increase frequency of confidence threshold updates (#147)
Previously, detector metadata (including confidence threshold) would be
cached for an hour, meaning updates to the confidence threshold via the
dashboard wouldn't be incorporated for a while. This was liable to
create confusing behavior, where a user adjusts the confidence threshold
but sees unexpected behavior on the edge for a while afterward. The
updated value would only be received once the cached value expired or
upon a re-deploy.

Now, cached detector metadata can only stick around for a maximum of 10
minutes, and otherwise will be re-computed once an image query is sent
in. Additionally, a background task runs after each image query request
is processed which checks how old the cached value for a detector's
metadata is. If it's older than 30 seconds, it will be refreshed (still
in the background, meaning it doesn't slow down query processing times -
other than occupying the worker for slightly longer).

Let me know if this seems like it's over-complicating things too much. I
tried to avoid defining a new cache class but wasn't able to figure out
a better way to do this. Definitely open to ideas though!