Gpu support on top of Balena w/ k3s (#125)
Support the edge-endpoint on top of k3s on top of balena on an x86
machine + nvidia GPU.

Doing this took a bit of work plus an assist in the form of some advice
from the folks at Balena.

There's a good amount of documentation in this PR about what is
happening, but the high level overview is this:
1. Follow this guide to get the `server` container to have GPU access
(`nvidia-smi`):
https://blog.balena.io/how-to-use-nvidia-gpu-on-x86-device-balenaos/
2. Install nvidia-container-toolkit on the `server` image and update the
containerd config.toml
3. Use k3s built in helm support to install the `nvidia-gpu-operator`.
Configure it to NOT install nvidia drivers or nvidia-container-toolkit
automatically

Additionally, to provide both GPU and continued CPU-only support, I
needed to create a custom deployment script `./deploy_balena.sh`. I
would LOVE suggestions for how to get rid of this or make it better.
More rationale is documented on the script. I'm also worried this might
conflict with glhub + edge-endpoint combined deployment.