Test with k3s revisited (#155)
This PR fixes this previous PR that was merged and then reverted:
https://github.com/groundlight/edge-endpoint/pull/142
The changes are largely the same with a few key differences:
1) The reason we reverted the PR -- there was a bug in setup edge
endpoint because of the changes I made to the kubernetes manifests that
relied on environment variables that we weren't defaulting correctly. I
fixed that
2) In order to avoid this mistake in the first place I added a test that
just runs setup-ee. While that script is run during our test-with-k3s
integration tests, we set some special environment variables for that
test specifically, which is the reason the bug wasn't caught in the
first PR. So this just runs it as close to how a user would as possible.
3) I think test-with-k3s can be run before the PR is merged not just
after.
4) I had some issues with the default stoage class. I kept getting the
error on my PVC: 'local-sc' not found. So I deleted the reference to the
storage class entirely, which according to some k8s documentation I
found online:

> The default Storage Class is automatically created by Kubernetes when
you set up a new cluster. This Storage Class is used when a Persistent
Volume Claim (PVC) does not specify a particular Storage Class, ensuring
that storage resources are provisioned even if no specific Storage Class
is requested.

The default storage class it uses in that case is 'local-path'. I
verified that this still worked by creating a file in the persistent
volume, deleting the pod, and verifying that it was still there when a
new pod came online.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>
Co-authored-by: Tom Faulhaber <github1_com@infolace.com>
Co-authored-by: robotrapta <79607467+robotrapta@users.noreply.github.com>