Fix handling of human_review=ALWAYS (#164)
When motion detection was removed, the check for `human_review=ALWAYS`
was accidentally removed, meaning currently even when
`human_review=ALWAYS` the edge answer will be returned if it's confident
enough. This adds the check back in to make sure we handle that
parameter correctly.

There's a test for this behavior
[here](https://github.com/groundlight/edge-endpoint/blob/a33b0ae46da67062d2ba926cc51525f6e39ab5a8/test/api/test_image_queries.py#L190)
which doesn't seem to be working correctly since it's been passing this
whole time. After a bit of poking around I couldn't figure out how to
fix the current test, so I think it might make sense to just leave it
until my new live tests are added (which are actually how I found this
bug!).