Integration tests (#149)
These tests expand on the basic integration we had for k3s. They are a
bit unusual as they weave python and bash together. We test that we can
setup the inference pods, submit to edge and get a low confidence answer
at first, submit to edge and escalate to cloud to train the model, pull
the updated model back down to the cloud into a new inference pod, and
now make a confident edge prediction.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>
Co-authored-by: Tom Faulhaber <github1_com@infolace.com>