Automatically refresh the ECR login credentials (#159)
Cron-job that every 2 hours refreshes the ECR login token.

I put all the needed functionality into the `edge-endpoint` container,
which adds a few tools that weren't there before, so it's bigger. I also
gave that container authz to create/destroy secrets. We could isolate
this stuff more, but it would be more complex. The auth stuff is limited
to the namespace.

Tasks:
- [x] Get the refresh script into Dockerfile
- [x] Fully test

---------

Co-authored-by: Ubuntu <ubuntu@ip-10-45-0-52.us-west-2.compute.internal>
Co-authored-by: Tyler Romero <tyler@groundlight.ai>