Save OODD models to repository (#175)
First edge-endpoint PR to get OODD running on the edge! This PR makes
use of the new information provided through the fetch-model-urls
endpoint, and saves OODD models to the edge. It sets up a new model
repository format, where the primary edge model versions are saved under
`<detector_id>/primary` within the model repository, instead of just
`<detector_id>`. The new OODD models will be saved under
`<detector_id>/oodd`.

To switch existing edge-endpoints over to the new model repository
format, the edge will pull the latest model for initial set up, and
clear out the old model versions in the repository. This will happen
even if the cloud model is the same as the current edge model.

In this PR, nothing is done to use the downloaded OODD model. Future PRs
will spin up a separate inference pod for it, and combine the primary
and OODD predictions into a single prediction.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>