Perform OODD inference on the edge (#179)
This is the 2nd PR to get OODD inference running on the edge. The [first
PR](https://github.com/groundlight/edge-endpoint/pull/175) downloaded
the models to the edge and saved them in the model repository. This PR
takes those saved OODD models for deployed detectors, spins up an
inference pod for them alongside the primary inference pod, and runs
OODD inference in parallel with primary inference.

In this PR, the OODD inference result isn't used yet. A third PR will
take the inference results from the primary and OODD models and combine
them into a final edge prediction.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>