Updating documentation in CONFIGURING-DETECTORS.md (#181)
This readme had some outdated information regarding how image queries
are handled when the detector's inference pod is not online. I have
updated it. I have tested this functionality today, so my description
should be accurate.

One thing that is still not clear to me: if I do not configure my
detector in `edge-config.yaml`, what is the default behavior? A
description of the default behavior might be nice to add to that yaml
and/or this readme.