Audit confident edge predictions (#156)
When the edge makes a confident prediction there is now a chance
(defined by the `confident_audit_rate` parameter) that the query will be
sent to the cloud to be audited after the edge answer is returned. This
allows us to better evaluate the accuracy of the edge pipeline.

The escalated image query is submitted with metadata `is_edge_audit =
True`, which will cause it to be audited in the cloud regardless of the
cloud ML result.

This change is fully backwards-compatible because the new parameter will
assume the default value if not present in the config. I tested this by
switching to the new edge-endpoint image on an already-running edge
endpoint and verified that it still worked correctly.

The current default rate is 1e-5, meaning there is a 0.001% chance to
audit a confident query.

There is currently no way to disable auditing (other than disabling
cloud escalations entirely), but I could add that if we think that might
be desired.

The `post_image_query` method is getting huge and confusing so I think
we should move more of the code into helper functions, but I plan to do
that in a separate PR.