Very minimal EE metrics reported to cloud (#191)
This supports the collecting data for most minimal reporting:

- How many edge devices are running
- Are they active or idle?  (When was the last image-query?)

It also adds a minimal HTML status page if you try to open the EE's port
in a browser. This is just a placeholder, but will become useful soon
enough.

This also include https://github.com/groundlight/edge-endpoint/pull/208
to get the e2e tests to pass.

---------

Co-authored-by: Tom Faulhaber <tom@groundlight.ai>
Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>
Co-authored-by: CoreyEWood <coreywood@outlook.com>