Don't include `results` in metadata if it would pass the max allowed size (#225)
Previously, with ~7+ ROIs in the `results` dict, cloud escalations would
fail because the metadata exceeded the maximum allowed size. This
ensures we don't exceed the size and restricts the contents of the
`results` data if necessary.

The max size is defined
[here](https://github.com/groundlight/python-sdk/blob/67c8d291da33f7b42302fa49c08bc73745210f68/src/groundlight/client.py#L342)
in the SDK. Unfortunately the number is hard-coded in, so if it changes
in the SDK we'll have to manually update it here.