Make test workflow reusable to trigger edge tests for new inference server images (#222)
This PR splits out building the EE image and the helm k3s tests into
reusable workflows, which can be called outside of this repo. The goal
for this update is to be able to trigger the EE tests when we build a
new inference image. Since this is a public repo, this does mean that
the workflow could be called by any other repo, but I'm not really
concerned about that since it requires the secrets to be passed in.

`test-with-k3s-helm` validates that the inference deployments come
online and runs full integration tests using them, so I think it's the
right selection to test a new inference image. `test-with-k3s-setup-ee`
would do the same, but it wasn't immediately clear how to specify the
inference image, and since we're switching over to helm I think it's
fine to just run that one.

The tests on this PR pass and were run on the correct images. This
[zuuul PR](https://github.com/positronix-ai/zuuul/pull/4883) will
trigger these new workflows when we build a new inference image, and has
been tested on this branch. Some modifications will need to be made to
switch it to the main branch, and it will need to be merged after this
one goes in.