Integration test tweaks and temporarily remove model improvement assertion (#236)
Temporarily removes the checks for the model being more confident once
we fetch an updated binary, since those are currently not reliable
because of the OODD implementation.

This will help the tests pass for now, and later we can revisit how to
do this in a more reliable way. We already verify that a new inference
pod is created via kubernetes-based checks, so the test will still
verify the key behavior of pulling an updated model binary and deploying
new inference pods.

Also makes some small organizational tweaks - I left comments in the
code about these.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>