Validate API token and fail fast if bad (#235)
This PR introduces a pre-install hook to test the provided
`GROUNDLIGHT_API_TOKEN` and fail the install immediately if the token
doesn't work. The job will log the results so that users can figure out
more details about why the token didn't work.

I also added some troubleshooting around this to the README.