Log activity hour for activity metrics (#243)
When saving these metrics in the cloud, I'd like to have access to the
hour they're reporting for in addition to the time they're sent. This PR
adds that to the activity metrics.

I also adjusted the cron job scheduling, I think before it was possible
for the report job to run before the collect job.