Adding edge canary on laptop (#182)
This PR adds an Edge Endpoint canary that runs on a laptop.

Completed so far:
1. a script that installs a cron job to periodically runs a load test
and check he result
2. The script creates a heartbeat detector and a Groundlight alert for
that detector, if it doesn't already exist. The idea is at the end of
every successful load test, the script will submit one image query to
the heartbeat detector. It will be submitted to the cloud, not edge. If
the heartbeat detector doesn't receive an image query for a given period
of time, Groundlight will send an email alert.
3. This canary writes local logs like this:
```
Cron job started at Fri Feb 21 05:05:01 PM PST 2025
2025-02-21 17:05:02 - INFO - Heartbeat alert already exists. Skipping...
2025-02-21 17:05:04 - INFO - Starting laptop edge canary test. Submitting 1000 image queries to det_2sxWe4bhmSjcAeqNk2R9wc2xaIb...
2025-02-21 17:08:00 - INFO - Processed 1000 image queries in 175.91 seconds at 5.68 queries per second.
2025-02-21 17:08:00 - INFO - Finished with an average confidence of 0.95.
2025-02-21 17:08:00 - INFO - Laptop edge canary seems to be online and functioning properly. Submitting heartbeat...
2025-02-21 17:08:01 - INFO - Heartbeat submitted!
Cron job finished at Fri Feb 21 05:08:01 PM PST 2025
```

What remains to be done:
1. Figure out how `edge-endpoint` picks up new docker images. Is this
happening automatically? Probably need to add some GHA that uploads
images with a `canaries` or `staging` tag prior to pushing to `latest`.
Maybe that should be a separate PR?