Make OODD requests sequential (#251)
After benchmarking, I determined that it's actually faster to run the
primary and OODD inference requests sequentially, as opposed to in
parallel with asyncio. OODD runs pretty quickly, so it seems that the
setup time to gather the async results is more than what it takes to
just run OODD after primary inference. This PR makes that change, to
improve the edge-endpoint's framerate capability.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>