Escalation queue reading and writing classes (currently unused) (#245)
This PR introduces the initial pieces of the escalation queue, the
`QueueWriter` and `QueueReader` classes.

The idea of this initial PR is that it doesn't touch or otherwise affect
any existing functionality - once merged, the changes here will be
entirely unused. Future PRs will build on this and begin actually using
the code. This strategy should make changes easier to review and more
incremental.

The actual code is (hopefully) not overly long or complex, but there's a
lot of tricky interactions that we want to make sure we get right. To
help this, I wrote tests for all the behaviors and scenarios I could
think of, but please review the tests carefully and let me know if you
think of more cases to test!

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>
Co-authored-by: Tom Faulhaber <tom@groundlight.ai>