Make use of the `from_edge` and `done_processing` fields (take two) (#216)
Trying again to implement what [this PR
did](https://github.com/groundlight/edge-endpoint/pull/189), except
without requiring the client submitting to the edge endpoint to have an
up-to-date SDK version.

Any Result produced on the edge will have `from_edge=True`.
Additionally, results produced in the cloud will never have
`from_edge=True`, so this field can decisively be used to determine if a
result was produced on the edge.

An IQ will have `done_processing=True` when polling for an answer at a
later point (based on the query's ID) has no chance of producing a
better answer. This is the case when a query was answered on the edge
and was not escalated to the cloud.

To ensure that setting these fields does not break things for users
still on old SDK versions, I wrote a script to iterate through previous
versions of the Groundlight Python SDK and submit an image query to an
edge endpoint running this branch's code. I tested the previous 30
versions and all succeeded, going back to `0.15.0` (released on April 1,
2024). I included this rough testing script in case we want to use it in
the future.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>