Add Kubernetes memory limits to avoid OOM crashes (#256)
This PR prevents the OOM system crashes we have been seeing when a user
attempts to use too many detectors at the same time.

A Kubernetes eviction policy is added that will start evicting pods when
the system-wide available memory drops below a certain threshold. This
threshold is defined as either a percentage of available memory, or an
absolute number, whichever is smaller.

Additionally, memory "requests" were added for system-critical
containers (each of the four containers in the edge-endpoint pod) so
that they will not be evicted, which would cause the system to break.

I included `memory_pressure_test.py` to test these changes. The changes
are effective; I have never seen the Edge Endpoint crash while these
changes are in effect. However, it does occasionally slow down. Also,
since it evicts inference pods, the user may not get edge answers for
all of their detectors. I think this is acceptable, since these changes
are meant to be a fail safe that simply prevents the system from
crashing. We can add more sophisticated logic in the future that
intelligently loads and offloads inference pods based on usage,
user-defined settings, etc.