Add escalation queue functionality (#227)
The pieces to make the escalation queue actually function!

There's many unit tests designed to test the core functionality of the
new functions. I've also done manual testing to verify that a) normal
escalations work as they did before and b) the escalation queue fulfills
the basic promise of holding on to escalations while offline and
escalating them when it's online again.

Further testing is planned to figure out what the behavior is in more
complicated scenarios (intermittent connectivity, unstable connection,
etc.) but this should be safe to merge before.

Some of the values chosen are arbitrary (e.g., request timeout schedule
and max number of cached requests) and can be tweaked as needed in the
future based on how it runs in practice.

### Key files:

- In `image_queries.py`: logic for when we write to the escalation
queue.
- In `edge-deployment.yaml`: creating a new container that continually
reads from the queue.
- In `manage_reader.py`: code that the escalation queue container runs
to read from the queue and do the escalations. Tricky logic for when to
retry a failed escalation vs give up.
- In `test_escalation_queue.py`: many tests for the queue logic.
Hopefully covering most basic functionality but let me know if you see
things that are missing.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>