Update namespace used for kubernetes stats (#278)
We have seen some 403 errors in opensearch logs for edge metrics, as
well as on the status page. I determined that those errors only occurred
on edge-endpoints set up with `setup-ee`, since the namespace used is
different.

I manually verified that I don't see 403 errors (or other errors) when
accessing the status page for an edge endpoint deployed with helm or one
deployed with `setup-ee` after these changes.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>