Distributing inference pods across available GPUs (#270)
This PR allows machines with multiple GPUs to utilize all of their GPUs.

I have updated the inference image to contain a script for selecting a
GPU, if there are multiple GPUs.

I also refactored the memory pressure script so that it could be used
for the sort of testing I was doing on this PR.