Manual network resilience testing with toxiproxy (#271)
Adds scripts that enable deploying a Toxiproxy instance within the same
kubernetes namespace as the edge deployment and routing
`api.groundlight.ai` traffic through it. See the README for info on the
available functionality.

Does not affect any of the default edge endpoint behavior. Provided
scripts are only used for manual testing.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>