Improve behavior of metadata refresh when experiencing network connectivity issues (#291)
This PR makes changes that cause the `edge-endpoint` container workers
to spend less time trying to refresh the detector metadata when there
are network connectivity issues. This ensures most of their time can be
spent responding to image query requests.

- Updates the `poetry.lock` file to pin `groundlight` to v0.23.2.
Previously it was on v0.23.0 (and the `pyproject.toml` already specifies
`>=0.23.0, <0.24.0`).

- Modifies the `TimestampedCache` defined in `utils.py` to update a
restored value's timestamp to the time of restoration, rather than
preserving the original timestamp.
- This means that when we restore a value after failing to update the
metadata, we won't try to refresh it again on the next request. This is
good for network outage scenarios to reduce the number of failed refresh
attempts we have.
  - I updated the tests to reflect this functionality change.
- Modifies the http transport retry behavior of the `Groundlight`
instance defined in `app_state.py`, which is utilized by the
`edge-endpoint` container.
- This utilizes a new parameter added to the Groundlight SDK in v0.23.2.
- Now, requests made through this instance that use idempotent methods
(i.e., not POST or PATCH) will not retry at all (down from the default
of 3 total retries).
- This helps ensure we don't stall too long when there are network
connectivity issues, avoiding holding up the worker as much as possible.
- I believe that currently the only affected request will be the
`get_detector` request made when refreshing detector metadata.
- Specifies a lower request timeout for the `get_detector` request when
refreshing detector metadata
- This utilizes a new parameter added to the Groundlight SDK in v0.23.2.
- Similarly to above, this ensures we fail more quickly when there are
network issues.
- I set it to `connect_timeout=2`, `read_timeout=3` - LMK if that
doesn't seem appropriate
- Very small addition to the Toxiproxy README (unrelated to other
changes)

I manually tested the changes to check that they a) don't break existing
functionality and b) address the issue they're supposed to solve:
- I confirmed that normal metadata refreshing still works fine, which it
does.
- I applied the same Toxiproxy disruption as during my previous testing,
when I saw the metadata refresh taking 40 seconds total to fail (four
tries with 10 seconds of waiting each). With the changes in this PR, it
now fails within 3 seconds (one try with max 3 seconds of waiting).
- I also verified that the cache timestamp change works as intended to
slow the rate of refresh attempts.