Add Bounding Box mode capability (#240)
This PR gets bounding box mode onto the edge!

My least favorite part of this is that we have to duplicate the
confidence calculation, since bbox mode uses multipredictions. For other
modes which use multipredictions, the confidence calculation is just
taking the max of the probabilities (which we also do on the edge side),
but the calculation is slightly more complex for bounding box mode. We
do a similar duplication for the OODD confidence calculation, so while
unfortunate, this isn't unprecedented.

---------

Co-authored-by: Auto-format Bot <autoformatbot@groundlight.ai>